refactor: 💡 Extract family tree from IDesign into its own files
