+++ libs/interfaces/src/lib/design/design.interface.ts
@@ -1,6 +1,5 @@
-import { ComponentRef } from '@angular/core';
-import { BoxDesignEnum, TreeDesignEnum } from '@assets';
 import { DesignTypeEnum, IBase, IUser } from '..';
+import { IFamilyTree } from './family-tree';
 
 export interface IDesign extends IBase {
   designId: string;
@@ -8,45 +7,4 @@ export interface IDesign extends IBase {
   designType: DesignTypeEnum;
   user: IUser;
   mutable: boolean;
-}
-export interface IFamilyTree {
-  font: FamilyTreeFontEnum;
-  backgroundTreeDesign: TreeDesignEnum;
-  boxSize: number;
-  banner: IFamilyTreeBanner;
-  boxes: IDraggableBox[];
-}
-
-export interface IDraggableBox {
-  x: number;
-  y: number;
-  previousX: number;
-  previousY: number;
-  dragging: boolean;
-  boxDesign: BoxDesignEnum;
-  // can't declare the actual component so the ref is any
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  inputRef?: ComponentRef<any>;
-  text: string;
-}
-
-export enum FamilyTreeFontEnum {
-  argestadisplay = 'argestadisplay',
-  argestadisplayItalic = 'argestadisplay-italic',
-  bairolBold = 'bairol-bold',
-  bairolBoldItalic = 'bairol-bold-italic',
-  calendasItalic = 'calendas-italic',
-  sansita = 'sansita',
-  roboto = 'roboto',
-  georgia = 'georgia',
-  spectral = 'spectral',
-  archaLight = 'archia-light',
-  archaMedium = 'archia-medium',
-  calendas = 'calendas',
-  knile = 'knile',
-}
-
-export interface IFamilyTreeBanner {
-  text: string;
-  style: 'first';
 }
