+++ libs/interfaces/src/lib/design/family-tree/family-tree-banner.interface.ts
@@ -1,0 +1,4 @@
+export interface IFamilyTreeBanner {
+  text: string;
+  style: 'first';
+}
