+++ libs/interfaces/src/lib/design/family-tree/family-tree.interface.ts
@@ -1,0 +1,10 @@
+import { TreeDesignEnum } from '@assets';
+import { FamilyTreeFontEnum, IDraggableBox, IFamilyTreeBanner } from '.';
+
+export interface IFamilyTree {
+  font: FamilyTreeFontEnum;
+  backgroundTreeDesign: TreeDesignEnum;
+  boxSize: number;
+  banner: IFamilyTreeBanner;
+  boxes: IDraggableBox[];
+}
