+++ libs/interfaces/src/lib/design/family-tree/index.ts
@@ -1,0 +1,4 @@
+export * from './draggable-box.interface';
+export * from './family-tree-banner.interface';
+export * from './family-tree-font.enum';
+export * from './family-tree.interface';
