+++ libs/interfaces/src/lib/design/index.ts
@@ -2,3 +2,4 @@ export * from './design-dimension.enum';
 export * from './design-type.enum';
 export * from './design.dto';
 export * from './design.interface';
+export * from './family-tree';
